+++ partials/mobile/mobile-search.php
@@ -14,22 +14,19 @@ if ( ! defined( 'ABSPATH' ) ) {
 $post_type = get_theme_mod( 'ocean_menu_search_source', 'any' );
 
 // Assign mobile search form unique ID.
-$ocean_msf_id = oceanwp_unique_id( 'ocean-mobile-search' );
+$ocean_msf_id = oceanwp_unique_id( 'ocean-mobile-search-' );
 $mosf_id      = esc_attr( $ocean_msf_id );
 ?>
 
 <div id="mobile-menu-search" class="clr">
-	<form method="get" action="<?php echo esc_url( home_url( '/' ) ); ?>" class="mobile-searchform" role="search" aria-label="<?php oceanwp_theme_strings( 'owp-string-mobile-search', 'oceanwp' ); ?>">
-		<label for="<?php echo $mosf_id; ?>">
-			<span class="screen-reader-text"><?php esc_html_e( 'Search for:', 'oceanwp' ); ?></span>
-			<input class="field" id="<?php echo $mosf_id; ?>" type="search" name="s" autocomplete="off" placeholder="<?php oceanwp_theme_strings( 'owp-string-mobile-search-text', 'oceanwp' ); ?>" />
-			<button type="submit" class="searchform-submit" aria-label="<?php oceanwp_theme_strings( 'owp-string-mobile-submit-search', 'oceanwp' ); ?>">
-				<?php oceanwp_icon( 'search' ); ?>
-			</button>
-			<?php if ( 'any' !== $post_type ) { ?>
-				<input type="hidden" name="post_type" value="<?php echo esc_attr( $post_type ); ?>">
-			<?php } ?>
-		</label>
+	<form aria-label="<?php oceanwp_theme_strings( 'owp-string-search-form-label' ); ?>" method="get" action="<?php echo esc_url( home_url( '/' ) ); ?>" class="mobile-searchform" role="search">
+		<input aria-label="<?php oceanwp_theme_strings( 'owp-string-search-field' ); ?>" value="" class="field" id="<?php echo $mosf_id; ?>" type="search" name="s" autocomplete="off" placeholder="<?php oceanwp_theme_strings( 'owp-string-mobile-search-text' ); ?>" />
+		<button aria-label="<?php oceanwp_theme_strings( 'owp-string-mobile-submit-search' ); ?>" type="submit" class="searchform-submit">
+			<?php oceanwp_icon( 'search' ); ?>
+		</button>
+		<?php if ( 'any' !== $post_type ) { ?>
+			<input type="hidden" name="post_type" value="<?php echo esc_attr( $post_type ); ?>">
+		<?php } ?>
 		<?php do_action( 'wpml_add_language_form_field' ); ?>
 	</form>
 </div><!-- .mobile-menu-search -->