+++ partials/scroll-top.php
@@ -23,4 +23,4 @@ $arrow = $arrow ? $arrow : 'angle_up';
 $position = apply_filters( 'ocean_scroll_top_position', get_theme_mod( 'ocean_scroll_top_position' ) );
 $position = $position ? $position : 'right'; ?>
 
-<a id="scroll-top" class="scroll-top-<?php echo esc_attr( $position ); ?>" href="#"><?php oceanwp_icon( $arrow ); ?></a>
+<a aria-label="<?php oceanwp_theme_strings( 'owp-string-scroll-top' ); ?>" href="#" id="scroll-top" class="scroll-top-<?php echo esc_attr( $position ); ?>"><?php oceanwp_icon( $arrow ); ?></a>